Move all CSV validation into ValidatedCSV
The main BulkProcessor class had too much knowledge of CSV validation.
Move all of that logic into HeaderValidator and rename it to
something more meaningful.

Also, we no longer need to catch the bizarre NoMethodError with the
specific message. That was a bug in older versions of ruby, carried
forward from Boardwalk. We no longer support the affected ruby
versions.