Refactor ItemProcessor and Handler
Break ItemProcessor into the CSVProcessor and RowProcessor roles.
A class implementing CSVProcessor takes in the entire CSV (represented
as an Array of Hashes) and processes it with the `#start` method.

A base class (BulkProcessor::CSVProcessor) is provided to build for a
common use case. This class implements the start method, which will
iterate over each row with a RowProcessor. When all rows have been
processed, it passes the results to a Handler role.

See the README or spec/support/roles from more details on these roles