Add code to partition a file into N files
There are 2 mechanisms to do this: Balanced, which breaks the input CSV into
N number of smaller CSVs all having the same row count (+/- 1 row). Boundary,
which attempts to balance the partitioned files, but will not break 2 consecutive
rows that have the same value for a given column. The partitioned files are
persisted to S3.